+++ spring-cloud-tencent-polaris-context/src/main/java/com/tencent/cloud/polaris/context/config/PolarisContextAutoConfiguration.java
@@ -23,16 +23,10 @@ import java.util.List;
 import com.tencent.cloud.polaris.context.ConditionalOnPolarisEnabled;
 import com.tencent.cloud.polaris.context.ModifyAddress;
 import com.tencent.cloud.polaris.context.PolarisConfigModifier;
+import com.tencent.cloud.polaris.context.PolarisSDKContextManager;
 import com.tencent.cloud.polaris.context.ServiceRuleManager;
-import com.tencent.polaris.api.core.ConsumerAPI;
-import com.tencent.polaris.api.core.ProviderAPI;
 import com.tencent.polaris.api.exception.PolarisException;
-import com.tencent.polaris.assembly.api.AssemblyAPI;
-import com.tencent.polaris.assembly.factory.AssemblyAPIFactory;
 import com.tencent.polaris.client.api.SDKContext;
-import com.tencent.polaris.factory.api.DiscoveryAPIFactory;
-import com.tencent.polaris.factory.api.RouterAPIFactory;
-import com.tencent.polaris.router.api.core.RouterAPI;
 
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
@@ -48,36 +42,10 @@ import org.springframework.core.env.Environment;
 @EnableConfigurationProperties({PolarisContextProperties.class})
 public class PolarisContextAutoConfiguration {
 
-	@Bean(name = "polarisContext", initMethod = "init", destroyMethod = "destroy")
+	@Bean(initMethod = "init")
 	@ConditionalOnMissingBean
-	public SDKContext polarisContext(PolarisContextProperties properties, Environment environment, List<PolarisConfigModifier> modifierList) throws PolarisException {
-		return SDKContext.initContextByConfig(properties.configuration(modifierList, () -> {
-			return environment.getProperty("spring.cloud.client.ip-address");
-		}, () -> {
-			return environment.getProperty("spring.cloud.polaris.local-port", Integer.class, 0);
-		}));
-	}
-
-	@Bean
-	@ConditionalOnMissingBean
-	public ProviderAPI polarisProvider(SDKContext polarisContext) throws PolarisException {
-		return DiscoveryAPIFactory.createProviderAPIByContext(polarisContext);
-	}
-
-	@Bean
-	@ConditionalOnMissingBean
-	public ConsumerAPI polarisConsumer(SDKContext polarisContext) throws PolarisException {
-		return DiscoveryAPIFactory.createConsumerAPIByContext(polarisContext);
-	}
-
-	@Bean
-	public RouterAPI polarisRouter(SDKContext polarisContext) throws PolarisException {
-		return RouterAPIFactory.createRouterAPIByContext(polarisContext);
-	}
-
-	@Bean
-	public AssemblyAPI assemblyAPI(SDKContext polarisContext) throws PolarisException {
-		return AssemblyAPIFactory.createAssemblyAPIByContext(polarisContext);
+	public PolarisSDKContextManager polarisSDKContextManager(PolarisContextProperties properties, Environment environment, List<PolarisConfigModifier> modifierList) throws PolarisException {
+		return new PolarisSDKContextManager(properties, environment, modifierList);
 	}
 
 	@Bean
@@ -87,7 +55,7 @@ public class PolarisContextAutoConfiguration {
 	}
 
 	@Bean
-	public ServiceRuleManager serviceRuleManager(SDKContext sdkContext, ConsumerAPI consumerAPI) {
-		return new ServiceRuleManager(sdkContext, consumerAPI);
+	public ServiceRuleManager serviceRuleManager(PolarisSDKContextManager polarisSDKContextManager) {
+		return new ServiceRuleManager(polarisSDKContextManager.getSDKContext(), polarisSDKContextManager.getConsumerAPI());
 	}
 }