+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - [feat:Remove error log from `DecodeTransferMetadataReactiveFilter`.](https://github.com/Tencent/spring-cloud-tencent/pull/989)
 - [Feature: add AssemblyFlow to support tsf.](https://github.com/Tencent/spring-cloud-tencent/pull/993)
 - [Refactoring: Refactor Circuitbreaker ut.](https://github.com/Tencent/spring-cloud-tencent/pull/997)
+- [feat:refactor SDKContext as static.](https://github.com/Tencent/spring-cloud-tencent/pull/1001)